requirements.txt
Flask==2.3.2
Flask_SSLify==0.1.5
mysql_connector_repackaged==0.3.1
Requests==2.31.0
